fastapi-api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
databases[postgresql]
passlib[bcrypt]
python-jose[cryptography]
pydantic
authtools
pytest
httpx